+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.6.7] - 2026-03-06
 - fix #81, persistent I2C address.
+- add **void resetPOR()** read back from non volatile RAM.
+- add **setConfiguration()** // replaces setConfigure in 0.7.0
+- add **getConfiguration()** // idem
+- remove "older experimental tags"
+- update examples (configuration calls)
 - update readme.md
 - update GitHub actions
 - minor edits